Renamed Header to AppBar
AppBar better reflects what this components is

- Renamed Header to AppBar
- Created new Header for admin interface
