+++ app/admin/page.tsx
@@ -1,11 +1,11 @@
 import React from "react";
-import Header from "@/components/admin/Header";
 import Sidebar from "@/components/admin/Sidebar";
+import AppBar from "@/components/admin/AppBar";
 
 function AdminPage() {
   return (
     <div className="min-h-screen bg-gray-50">
-      <Header />
+      <AppBar />
       <div className="flex pt-14">
         {/* Sidebar */}
         <Sidebar />
